chore(deps/runner): upgrade base image to follow "latest" tag
The base-image runners seem very stable, and we more often than not run
into errors and issues due to our base image being too old, rather than
it has been updated. Hence I'm changing it to follow the "latest" tag,
which is used for the latest versioned release of the base image.